Example health pickup
Created an effect actor class that acts as a pick up and hacks the player's health to increase.
Uses const_cast to work around not using a gameplay effect.

Should only be used during early exploration of an idea whilst figuring out proper functionality, later to be replaced by a GE solution.