Recreated GameplayTags in C++ (lesson 88)
Deviated from lesson plan as I used the UE Macro for adding tags, which saves a bit of effort. Not sure if I did it the right way, but they work.

Also, I created health and mana tags this way, and in the lesson they were left in the original state of being manually entered in the editor.